create sui-package-resolver crate
