+++ crates/sui-package-resolver/fixtures/packages/a0/sources/n.move
@@ -1,0 +1,10 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+#[allow(unused_field)]
+module a::n {
+    struct T0 {
+        t: a::m::T1<u16, u32>,
+        u: a::m::T2,
+    }
+}
