+++ crates/sui-package-resolver/fixtures/packages/a1/Move.toml
@@ -1,0 +1,7 @@
+[package]
+name = "A"
+version = "0.0.1"
+published-at = "0xa1"
+
+[addresses]
+a = "0xa0"
