+++ crates/sui-package-resolver/fixtures/packages/b0/Move.toml
@@ -1,0 +1,10 @@
+[package]
+name = "B"
+version = "0.0.1"
+published-at = "0xb0"
+
+[dependencies]
+A = { local = "../a0" }
+
+[addresses]
+b = "0xb0"
