+++ crates/sui-package-resolver/fixtures/packages/s1/sources/m.move
@@ -1,0 +1,13 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+#[allow(unused_field)]
+module s::m {
+    struct T0 {
+        x: u64
+    }
+
+    struct T1 {
+        x: u256
+    }
+}
